Move Details and AlignmentStatistics to separate header file
