+++ src/aln.hpp
@@ -10,61 +10,8 @@
 #include "sam.hpp"
 #include "aligner.hpp"
 #include "insertsizedistribution.hpp"
+#include "statistics.hpp"
 
-struct AlignmentStatistics {
-    std::chrono::duration<double> tot_read_file{0};
-    std::chrono::duration<double> tot_construct_strobemers{0};
-    std::chrono::duration<double> tot_find_nams{0};
-    std::chrono::duration<double> tot_time_rescue{0};
-    std::chrono::duration<double> tot_find_nams_alt{0};
-    std::chrono::duration<double> tot_sort_nams{0};
-    std::chrono::duration<double> tot_extend{0};
-    std::chrono::duration<double> tot_write_file{0};
-
-    uint64_t n_reads{0};
-    uint64_t n_hits{0}; // non-rescue hits
-    uint64_t n_rescue_hits{0};
-    uint64_t n_nams{0};
-    uint64_t n_rescue_nams{0};
-    uint64_t tot_aligner_calls{0};
-    uint64_t tot_rescued{0};
-    uint64_t tot_all_tried{0};
-    uint64_t inconsistent_nams{0};
-    uint64_t nam_rescue{0};
-
-    AlignmentStatistics operator+=(const AlignmentStatistics& other) {
-        this->tot_read_file += other.tot_read_file;
-        this->tot_construct_strobemers += other.tot_construct_strobemers;
-        this->tot_find_nams += other.tot_find_nams;
-        this->tot_time_rescue += other.tot_time_rescue;
-        this->tot_find_nams_alt += other.tot_find_nams_alt;
-        this->tot_sort_nams += other.tot_sort_nams;
-        this->tot_extend += other.tot_extend;
-        this->tot_write_file += other.tot_write_file;
-        this->n_reads += other.n_reads;
-        this->n_hits += other.n_hits;
-        this->n_rescue_hits += other.n_rescue_hits;
-        this->n_nams += other.n_nams;
-        this->n_rescue_nams += other.n_rescue_nams;
-        this->tot_aligner_calls += other.tot_aligner_calls;
-        this->tot_rescued += other.tot_rescued;
-        this->tot_all_tried += other.tot_all_tried;
-        this->inconsistent_nams += other.inconsistent_nams;
-        this->nam_rescue += other.nam_rescue;
-        return *this;
-    }
-
-    AlignmentStatistics operator+=(const Details& details) {
-        this->n_nams += details.nams;
-        this->n_rescue_nams += details.rescue_nams;
-        this->nam_rescue += details.nam_rescue;
-        this->tot_rescued += details.mate_rescue;
-        this->tot_all_tried += details.tried_alignment;
-        this->inconsistent_nams += details.nam_inconsistent;
-
-        return *this;
-    }
-};
 
 enum class OutputFormat {
     SAM,
