+++ src/sam.hpp
@@ -6,6 +6,7 @@
 #include "kseq++/kseq++.hpp"
 #include "refs.hpp"
 #include "cigar.hpp"
+#include "statistics.hpp"
 
 
 struct Alignment {
@@ -41,30 +42,6 @@ enum SamFlags {
 enum struct CigarOps {
     EQX = 0,  // use = and X CIGAR operations
     M = 1,    // use M CIGAR operations
-};
-
-/* Details about aligning a single or paired-end read */
-struct Details {
-    bool nam_rescue{false}; // find_nams_rescue() was needed
-    uint32_t nams{0};  // No. of NAMs found
-    uint32_t rescue_nams{0}; // No. of NAMs found during rescue
-    uint32_t nam_inconsistent{0};
-    uint32_t mate_rescue{0}; // No. of times rescue by local alignment was attempted
-    uint32_t tried_alignment{0}; // No. of computed alignments (get_alignment or rescue_mate)
-    uint32_t gapped{0};  // No. of gapped alignments computed (in get_alignment)
-    uint32_t best_alignments{0}; // No. of best alignments with same score
-
-    Details& operator+=(const Details& other) {
-        nam_rescue = nam_rescue || other.nam_rescue;
-        nams += other.nams;
-        rescue_nams += other.rescue_nams;
-        nam_inconsistent += other.nam_inconsistent;
-        mate_rescue += other.mate_rescue;
-        tried_alignment += other.tried_alignment;
-        gapped += other.gapped;
-        best_alignments += other.best_alignments;
-        return *this;
-    }
 };
 
 class Sam {
